+++ shpsync/core/project_handler.py
