+++ shpsync/core/qgis_utils.py
