+++ shpsync/gui/icon.png
